edited apple_division_test

+++ tests/01_Introductory/apple_division_test.cpp
@@ -5,18 +5,18 @@ long long solve_apple_division(int n, const std::vector<long long> &P);
 
 TEST(Introductory, AppleDivision_BaseCase)
 {
-    long long expected = 1;
+    int expected = 1;
     EXPECT_EQ(solve_apple_division(5, std::vector<long long>{3, 2, 7, 4, 1}), expected);
 }
 
 TEST(Introductory, AppleDivision_EdgeCaseBig)
 {
-    long long expected = 88;
+    int expected = 88;
     EXPECT_EQ(solve_apple_division(5, std::vector<long long>{3, 2, 7, 100}), expected);
 }
 
 TEST(Introductory, AppleDivision_EdgeCaseOne)
 {
-    long long expected = 5;
+    int expected = 5;
     EXPECT_EQ(solve_apple_division(1, std::vector<long long>{5}), expected);
 }
